requirements.txt
baidu-aip>=2.2.18.0
Pillow>=9.0.0
requests>=2.25.1
chardet>=3.0.4